Exchange rate time series

Time series plans are only available on "professional" and above plans. The maximum allowed length in a single request is 365 days.

Each currency symbol is displayed in its own column, with a value relative to the base currency on the given date.

Time series data is only available for "Professional", "Professional Plus" and "Enterprise" accounts.

fixer_time_series(start_date, end_date, base = "EUR", symbols = NULL)

Arguments

start_date

The start date of requested time series, in "YYYY-MM-DD" style, or any style convertable to "YYYY-MM-DD" using as.Date().

end_date

The end date of requested time series, in "YYYY-MM-DD" style, or any style convertable to "YYYY-MM-DD" using as.Date().

base

The base currency to index other currencies against. Defaults to "EUR". Other base currencies are only available on paid plans.

symbols

A character vector of the symbols of currencies to return exchange rates for, or a string for a single currency. Defaults to NULL and returns all available currencies. See fixer_symbols() for details on symbol options.

Value

A tibble with the exchange rate from the base currency to the given currency symbols for each date in the requested range.

See also

Examples

if (FALSE) {

x <- fixer_time_series(start_date = "2017-04-30", end_date = "2017-06-20",
                       base = "GBP", symbols = c("USD", "EUR"))

}