Historic exchange rates

Historical exchange rates are only available on paid fixer.io accounts.

fixer_historical(date = NULL, base = "EUR", symbols = NULL)

Arguments

date
A date in YYYY-MM-DD format, or any value that can be coerced to YYYY-MM-DD format with as.Date(). Defaults to NULL, which returns the latest conversion data.

base
The base currency to index other currencies against. Defaults to "EUR". Can only be changed with paid plans.

symbols
The symbols of currencies to return exchange rates for. Defaults to NULL and returns all available currencies. See fixer_symbols] for details on symbol options.

Value

A tibble with exchange rates to the base currency on a given date.

Examples

if (FALSE) {

historical <- fixer_historical(date = "2017-05-18")

}